requirements.txt
numpy == 1.26.3
matplotlib == 3.8.3
pillow == 10.2.0
tensorflow == 2.10.0
fastapi == 0.115.6
uvicorn == 0.33.0